+++ runtime/index.js
@@ -11,5 +11,7 @@ const contextDiscovery = require("./context-discovery");
 const actionPlanner = require("./action-planner");
 const promotionGuidance = require("./promotion-guidance");
 const workflowRouter = require("./workflow-router");
+const { classifyIntent } = require("./intent-classifier");
+const { resolveRoute, DEFAULT_ROUTING_TABLE, KNOWN_INTENTS } = require("./route-resolver");
 
-module.exports = { decide, score, append, journalPaths, ...policy, ...session, ...projectPolicy, ...contextDiscovery, ...actionPlanner, ...promotionGuidance, ...workflowRouter };
+module.exports = { decide, score, append, journalPaths, ...policy, ...session, ...projectPolicy, ...contextDiscovery, ...actionPlanner, ...promotionGuidance, ...workflowRouter, classifyIntent, resolveRoute, DEFAULT_ROUTING_TABLE, KNOWN_INTENTS };